AVRO-4327: [Java] Use a long map loop counter to avoid int overflow
CodeQL flagged a comparison of a narrow int loop counter against a wider
long block count in the fast reader's MapReader. A map block count above
Integer.MAX_VALUE would overflow the int counter and never satisfy the
loop condition. Use a long counter, matching the array reader in the same
file.

+++ lang/java/avro/src/main/java/org/apache/avro/io/FastReaderBuilder.java
@@ -684,7 +684,7 @@ public class FastReaderBuilder {
         Map<Object, Object> targetMap = new HashMap<>();
 
         while (l > 0) {
-          for (int i = 0; i < l; i++) {
+          for (long i = 0; i < l; i++) {
             Object key = keyReader.read(null, decoder);
             Object value = valueReader.read(null, decoder);
             targetMap.put(key, value);
